Fixed library data size and formatting/parsing of large uint and signed numbers
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/837



git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@3571 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/extra/ustdio/ufmt_cmn.c
@@ -45,13 +45,13 @@ ufmt_isdigit(UChar     c,
 void 
 ufmt_ltou(UChar     *buffer, 
       int32_t     *len,
-      long         value, 
-      int32_t     radix,
+      uint32_t         value, 
+      uint32_t     radix,
       UBool    uselower,
       int32_t    minDigits)
 {
   int32_t     length     = 0;
-  long         q;
+  uint32_t         q;
   int8_t     digit;
   UChar    *left, *right, temp;
